fix(multiuser): Phase 3.3 — P0 stopwatch read leak + pause/resume + UX
P0-A: Stopwatch read leak — every StopwatchManager method used a static
"user_primary" Redis key for all users. User 2's active timer was visible
in user 1's /stopwatch/status response. Fix: _user_key() resolves the
numeric user_id from the scoping ContextVar and uses it as the Redis
namespace. All public methods (start, pause, resume, stop, get_status,
check_early_stop, correct_readiness) now derive the key per-request.
Adversarial test #13 added: user 99 starts timer, user 98 sees active=false.
Verified via curl: user 2 active, user 1 inactive.

P0-B: Pause/resume 500 — Phase 3.2 converted total_paused_minutes to
Float in the DB but StopwatchResumeResponse and StopwatchStatusResponse
still declared int fields. Pydantic rejected fractional values. Fix:
paused_minutes and total_paused_minutes now float in response schemas.

P0-C: completion_pct accepted 500% — HTML min/max only constrains spinner
arrows, not typed values. Fix: onChange handler now clamps to [0,100] in
real time so the field visually rejects out-of-range input. Backend
already had Field(ge=0, le=100).

P0-D: User 2 has 2 tasks — both legitimate (pre-fix "Test" from 4/9,
today's "Test 1"). No orphaned leak data.

P1-1: Error messages now show backend detail.message instead of raw JSON.
P1-3: Sort direction reversed — most recent tasks first (descending).
P1-4: Voided tasks filtered from Today view (frontend, voided_at exposed
in query response).
P1-5: Void affordance — trash icon on terminal tasks (EXECUTED/SKIPPED).
P1-6: Overdue auto-skip worker confirmed running, per-user scoped, 30min.

docs/hypotheses.md: H4 motivated underestimation hypothesis from Form A.

55 passed, tsc clean.

+++ backend/app/schemas/stopwatch.py
@@ -74,8 +74,8 @@ class StopwatchPauseResponse(BaseModel):
 
 class StopwatchResumeResponse(BaseModel):
     resumed: bool
-    paused_minutes: int
-    total_paused_minutes: int
+    paused_minutes: float
+    total_paused_minutes: float
 
 
 class StopwatchStatusResponse(BaseModel):
@@ -86,7 +86,7 @@ class StopwatchStatusResponse(BaseModel):
     start_time: Optional[datetime] = None
     elapsed_minutes: Optional[int] = None
     paused: bool = False
-    total_paused_minutes: int = 0
+    total_paused_minutes: float = 0
 
 
 class ReadinessCorrectionRequest(BaseModel):
